// electron-ui/node-metaverse/lib/classes/messages/CompletePingCheck.d.ts
import { MessageFlags } from '../../enums/MessageFlags';
import { MessageBase } from '../MessageBase';
import { Message } from '../../enums/Message';
export declare class CompletePingCheckMessage implements MessageBase {
    name: string;
    messageFlags: MessageFlags;
    id: Message;
    PingID: {
        PingID: number;
    };
    getSize(): number;
    writeToBuffer(buf: Buffer, pos: number): number;
    readFromBuffer(buf: Buffer, pos: number): number;
}
//# sourceMappingURL=CompletePingCheck.d.ts.map